orgcrawler-v2/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.6.3
cryptography==43.0.3
cssselect==1.3.0
distro==1.9.0
duckduckgo_search==7.4.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.14.0
h2==4.2.0
hf-xet==1.1.4
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.72.6.post1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.0
narwhals==1.43.0
networkx==3.3
nltk==3.9.1
numpy==2.1.2
ollama==0.4.7
openai==1.88.0
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
pillow==10.4.0
playwright==1.52.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pyOpenSSL==25.1.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
socksio==1.0.0
soupsieve==2.6
streamlit==1.45.1
streamlit-autorefresh==1.0.1
sympy==1.13.1
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
volcengine-python-sdk==3.0.14
watchdog==6.0.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
